rename correct files
